PawLIB 1.0rc1
Summary:
(NOTE) **UPDATE:** I am re-purposing this Differential for landing PawLIB 1.0rc1.

----

I am expanding the sorting algorithms for PawLIB, to offer the fastest known implementations of various algorithms.

In this initial commit, I am including...

- The Dual-Pivot Quicksort algorithm, with CSI comments.
- An insertion sort algorithm.
- A comparison test in `main.cpp` in Goldilocks, demonstrating the speed gains of the quicksort over selection sort.

Test Plan: Ensure all built-in tests run. (Will expand this later.)

Reviewers: sataylor

Reviewed By: sataylor

Subscribers: jgtheodore

Projects: #programming_dept, #pawlib_core_pawlib, #pawsort_pawlib, #pawstring_pawlib, #goldilocks_pawlib, #bss_team, #imf_team

Maniphest Tasks: T774, T773, T763, T762, T761, T758, T757

Differential Revision: https://hawksnest.ddns.net:8446/D122

+++ pawlib-source/src/binconv.cpp
@@ -152,7 +152,7 @@ float binconv::toFloat(string binary)
 {
     //the value starts at 0
     float val = 0;
-    if(binary.find("1") == -1)
+    if(binary.find("1") == (size_t) (-1))
         return 0;
     //used to determine if the number is a positive or negative number
     int theSign = 1;
@@ -295,7 +295,7 @@ double binconv::toDouble(string binary)
 {
     //stores the new decimal value
     double val = 0;
-    if(binary.find("1") == -1)
+    if(binary.find("1") == (size_t)(-1))
         return 0;
     int theSign = 1;
     //if the sign bit is set, the number needs to be multiplied by a negative 1 at the end
@@ -376,7 +376,7 @@ string binconv::binaryToString(string binary)
 {
     stringstream r;
     string character;
-    for(int i = 0; i < binary.length() / 8; i++)
+    for(unsigned int i = 0; i < binary.length() / 8; i++)
     {
         character = binary.substr(i * 8, 8);
         r << toChar(character);
